Add Form component: implement multi-question survey with API integration and user session handling
